Consolidate Transform-related code
